requirements.txt
gradio
transformers
torch
newspaper3k
PyMuPDF
bert-extractive-summarizer
